pyproject.toml
[project]
name = "maritime-kg"
version = "0.1.0"
description = "대화형 해사서비스 플랫폼을 위한 지식그래프 모델"
authors = [
    {name = "인사이트마이닝", email = "info@insightmining.io"},
]
keywords = ["maritime", "knowledge-graph", "neo4j", "kriso", "ontology", "nlp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
license = {text = "Proprietary"}
requires-python = ">=3.10"
readme = "README.md"
dependencies = [
    "neo4j~=5.26.0",
    "python-dotenv~=1.0.1",
    "langchain~=0.3.14",
    "langchain-neo4j~=0.5.0",
    "langchain-ollama~=0.3.0",
    "requests~=2.32.3",
    "beautifulsoup4~=4.12.3",
    "fastapi~=0.109.0",
    "uvicorn[standard]~=0.34.0",
    "pydantic>=2.0,<3.0",
    "neo4j-graphrag[ollama]~=1.13.0",
    "PyJWT>=2.0.0,<3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.4.0",
    "mypy>=1.8.0",
    "httpx>=0.25.0",
]

[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["core", "domains"]
include = ["kg*", "maritime*"]

[tool.pytest.ini_options]
pythonpath = [".", "core", "domains"]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers --tb=short"
markers = [
    "unit: Pure Python unit tests (no external dependencies)",
    "integration: Tests requiring Neo4j connection",
    "e2e: End-to-end scenario tests",
    "slow: Long-running tests (benchmarks, stress tests)",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501"]  # line length handled separately

[tool.ruff.lint.isort]
known-first-party = ["kg", "maritime"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
ignore_missing_imports = true

[tool.coverage.run]
source = ["kg", "maritime"]
omit = ["kg/crawlers/*", "maritime/crawlers/*", "maritime/schema/*", "maritime/rbac/*", "kg/utils/verification.py", "*/test_*"]

[tool.coverage.report]
show_missing = true
skip_covered = true
fail_under = 75
